localization

+++ app/Http/Controllers/API/OrderController.php
@@ -20,21 +20,25 @@ class OrderController extends Controller
     public function checkout(Request $request)
     {
         $user = Auth::user();
-        // TODO::check cart empty
-        $order = Order::create(['user_id' => $user->id, 'address_id' => $user->addresses->first()->id, 'subtotal' => $user->cart->total(), 'shipping' => 20, 'total' => $user->cart->total() + 20,]);
-        foreach ($user->cart->products as $product) {
-            OrderProduct::create(['product_id' => $product->product_id, 'order_id' => $order->id, 'quantity' => $product->quantity, 'total' => $product->product->price * $product->quantity]);
-            $product->product->update(['stock' => $product->product->stock - $product->quantity]);
+        if ($user->Cart->products->count() > 0) {
+            
+            $order = Order::create(['user_id' => $user->id, 'address_id' => $user->addresses->first()->id, 'subtotal' => $user->cart->total(), 'shipping' => 20, 'total' => $user->cart->total() + 20,]);
+            foreach ($user->cart->products as $product) {
+                OrderProduct::create(['product_id' => $product->product_id, 'order_id' => $order->id, 'quantity' => $product->quantity, 'total' => $product->product->price * $product->quantity]);
+                $product->product->update(['stock' => $product->product->stock - $product->quantity]);
+            }
+            $user->cart->products()->delete();
+            // dd($user->name);
+            $orderData = [
+                'order_id' => $order->id,
+                'user_name' => $user->name,
+                'address' => $user->addresses->first()->city,
+                'total' => $order->total
+            ];
+            OrderNotificationJob::dispatch($orderData);
+        } else {
+            return response()->json(['Cart is empty']);
         }
-        $user->cart->products()->delete();
-        // dd($user->name);
-        $orderData = [
-            'order_id' => $order->id,
-            'user_name' => $user->name,
-            'address' => $user->addresses->first()->city,
-            'total' => $order->total
-        ];
-        OrderNotificationJob::dispatch($orderData);
         return response()->json(['Total' => $order->total]);
     }
 }
